fix(combobox-web): remove state logic, manipulate path values on hover via css
