+++ packages/pluggableWidgets/combobox-web/src/ui/Combobox.scss
@@ -11,6 +11,7 @@ $unselected-item: #ffffff;
 $hover-color: #ced0d3;
 $disabled-color: #f8f8f8;
 $disabled-text-color: #6c7180;
+$select-all-button-fill: #24276c;
 .widget-combobox {
     flex-grow: 1;
     position: relative;
@@ -277,6 +278,28 @@ $disabled-text-color: #6c7180;
         }
         &.active {
             color: $brand-primary;
+            transition: color 0.2s ease-in-out;
+            &:hover {
+                transition: color 0.2s ease-in-out;
+                color: $brand-primary;
+                .select-all-icon-path-1 {
+                    d: path(
+                        "M1.95142 1.79123C2.09629 1.77861 2.24293 1.77217 2.39108 1.77217H13.2608C16.0222 1.77217 18.2608 4.01075 18.2608 6.77217V16.736C18.2608 17.22 18.192 17.688 18.0637 18.1307C19.2419 17.216 20 15.786 20 14.179V5C20 2.23858 17.7614 0 15 0H5.78613C4.24605 0 2.86861 0.696288 1.95142 1.79123Z"
+                    );
+                    fill: $select-all-button-fill;
+                }
+
+                rect {
+                    fill: $select-all-button-fill;
+                }
+
+                .select-all-icon-path-2 {
+                    d: path(
+                        "M12.0351 7.98986C11.8678 7.82369 11.6415 7.73042 11.4055 7.73042C11.1696 7.73042 10.9433 7.82369 10.7759 7.98986L8.72192 10.0419L6.67684 7.98986C6.50868 7.82186 6.2806 7.72748 6.04278 7.72748C5.80496 7.72748 5.57688 7.82186 5.40872 7.98986C5.24055 8.15786 5.14608 8.38571 5.14608 8.6233C5.14608 8.86089 5.24055 9.08874 5.40872 9.25674L7.46273 11.2998L5.40872 13.3429C5.31523 13.4229 5.2393 13.5213 5.1857 13.632C5.1321 13.7427 5.10197 13.8632 5.09722 13.9861C5.09247 14.109 5.1132 14.2315 5.15809 14.346C5.20299 14.4605 5.27109 14.5644 5.35813 14.6514C5.44516 14.7383 5.54924 14.8064 5.66384 14.8512C5.77844 14.8961 5.90108 14.9168 6.02407 14.912C6.14706 14.9073 6.26774 14.8772 6.37854 14.8236C6.48934 14.7701 6.58785 14.6942 6.66791 14.6008L8.72192 12.5578L10.767 14.6008C10.9378 14.747 11.1576 14.8234 11.3824 14.8147C11.6071 14.806 11.8203 14.713 11.9794 14.5541C12.1384 14.3952 12.2316 14.1822 12.2403 13.9576C12.249 13.7331 12.1725 13.5136 12.0262 13.3429L9.98112 11.2998L12.0262 9.25674C12.1105 9.17439 12.1776 9.07618 12.2237 8.96778C12.2698 8.85938 12.294 8.74294 12.2948 8.62516C12.2957 8.50738 12.2731 8.3906 12.2286 8.28157C12.184 8.17253 12.1183 8.07338 12.0351 7.98986Z"
+                    );
+                    fill: $unselected-item;
+                }
+            }
         }
     }
